Add MANIFEST.in.
Usage package data to bundle the README and LICENSE files does not
seem to work with either the zip_safe flag or source (tarball)
distributions.  Adding a MANIFEST.in fixes source dists and removing
the zip_safe=True makes it safe to access these files using a
straight open().read().