Fix AsyncTestCase.published_messages
The current implementation expects all calls to publish_message to be
made with kwargs. Calling publish_message with positional arguments or a
mix of the two results in an exception. For example, `KeyError: 'exchange'`
is raised when the call was made solely with positional arguments. Ex.

    self.publish_message('a', 'b', 'c', 'd')

The fixed implmentation simply expands both args and kwargs.

Tested in python 3.7 and 3.8